chore: test packages
