+++ internal/pkg/export/export_fuzz.go
@@ -1,6 +1,6 @@
 // -----------------------------------------------------------------------------
 // SPDX-FileCopyrightText: Copyright © 2024 bomctl a Series of LF Projects, LLC
-// SPDX-FileName: internal/pkg/merge/merge_fuzz.go
+// SPDX-FileName: internal/pkg/export/export_fuzz.go
 // SPDX-FileType: SOURCE
 // SPDX-License-Identifier: Apache-2.0
 // -----------------------------------------------------------------------------
@@ -17,27 +17,24 @@
 // limitations under the License.
 // -----------------------------------------------------------------------------
 
-package merge_test
+package export
 
 import (
 	"context"
 	"testing"
 
-	"github.com/bomctl/bomctl/internal/pkg/merge"
 	"github.com/bomctl/bomctl/internal/pkg/options"
 )
 
-const (
-	id1 = "urn:uuid:f360ad8b-dc41-4256-afed-337a04dff5db"
-	id2 = "urn:uuid:f360ad8b-dc41-4256-afed-337a04dff5dc"
-)
+const testID = "urn:uuid:f360ad8b-dc41-4256-afed-337a04dff5db"
 
-func FuzzMerge(f *testing.F) {
-	f.Add(id1, id2)
-	f.Fuzz(func(t *testing.T, id1, id2 string) {
-		_, err := merge.Merge(
-			[]string{id1, id2},
-			&options.MergeOptions{Options: options.New().WithContext(context.Background())},
+func FuzzExport(f *testing.F) {
+	f.Add(testID)
+
+	f.Fuzz(func(t *testing.T, id string) {
+		err := Export(
+			id,
+			&options.ExportOptions{Options: options.New().WithContext(context.Background())},
 		)
 
 		if err == nil {
