+++ internal/pkg/fetch/fetch_fuzz.go
@@ -1,6 +1,6 @@
 // -----------------------------------------------------------------------------
 // SPDX-FileCopyrightText: Copyright © 2024 bomctl a Series of LF Projects, LLC
-// SPDX-FileName: internal/pkg/export/export_fuzz.go
+// SPDX-FileName: internal/pkg/fetch/fetch_fuzz.go
 // SPDX-FileType: SOURCE
 // SPDX-License-Identifier: Apache-2.0
 // -----------------------------------------------------------------------------
@@ -17,25 +17,24 @@
 // limitations under the License.
 // -----------------------------------------------------------------------------
 
-package export_test
+package fetch
 
 import (
 	"context"
 	"testing"
 
-	"github.com/bomctl/bomctl/internal/pkg/export"
 	"github.com/bomctl/bomctl/internal/pkg/options"
 )
 
-const testID = "urn:uuid:f360ad8b-dc41-4256-afed-337a04dff5db"
+const testURL = "https://raw.githubusercontent.com/bomctl/bomctl-playground/main/bomctl_bomctl_v0.3.0.cdx.json"
 
-func FuzzExport(f *testing.F) {
-	f.Add(testID)
+func FuzzFetch(f *testing.F) {
+	f.Add([]byte(testURL))
 
-	f.Fuzz(func(t *testing.T, id string) {
-		err := export.Export(
-			id,
-			&options.ExportOptions{Options: options.New().WithContext(context.Background())},
+	f.Fuzz(func(t *testing.T, data []byte) {
+		_, err := Fetch(
+			string(data),
+			&options.FetchOptions{Options: options.New().WithContext(context.Background())},
 		)
 
 		if err == nil {
